Fixed assertion at EventQueue::addAnEvent
When running 8000 neurons static full connection model, an assertion
occurred at EventQueue::addAnEvent. This is because an integer overflow
happened at AllSpikingSynapses::setupSynapses.

+++ Synapses/AllSpikingSynapses.cpp
@@ -81,8 +81,10 @@ void AllSpikingSynapses::setupSynapses(const int num_neurons, const int max_syna
         // create a pre synapse spike queue & initialize it
         preSpikeQueue = new EventQueue();
 #if defined(USE_GPU)
-        int nMaxInterClustersOutgoingEvents = max_total_synapses * sim_info->maxFiringRate * sim_info->deltaT * sim_info->minSynapticTransDelay;
-        int nMaxInterClustersIncomingEvents = max_total_synapses * sim_info->maxFiringRate * sim_info->deltaT * sim_info->minSynapticTransDelay;
+        // max_total_synapses * sim_info->maxFiringRate may overflow the maximum range
+        // of 32 bits integer, so we cast it uint64_t
+        int nMaxInterClustersOutgoingEvents = (uint64_t) max_total_synapses * sim_info->maxFiringRate * sim_info->deltaT * sim_info->minSynapticTransDelay;
+        int nMaxInterClustersIncomingEvents = (uint64_t) max_total_synapses * sim_info->maxFiringRate * sim_info->deltaT * sim_info->minSynapticTransDelay;
 
         // initializes the pre synapse spike queue
         preSpikeQueue->initEventQueue(clr_info->clusterID, max_total_synapses, nMaxInterClustersOutgoingEvents, nMaxInterClustersIncomingEvents);
